Use the new Type enum and move GValue to the glib crate
