+++ src/gtk/widgets/tree_store.rs
@@ -13,25 +13,27 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with rgtk.  If not, see <http://www.gnu.org/licenses/>.
 
+use glib::{to_bool, GValue, Type};
+use glib::translate::ToGlib;
 use gtk::{self, ffi};
-use glib::ffi::GType;
 use gtk::TreeIter;
 use std::ffi::CString;
 use std::num::ToPrimitive;
-use glib::to_bool;
 
 pub struct TreeStore {
     pointer: *mut ffi::C_GtkTreeStore
 }
 
 impl TreeStore {
-    pub fn new(column_types: &[GType]) -> Option<TreeStore> {
-        let tmp_pointer = unsafe { ffi::gtk_tree_store_newv(column_types.len().to_i32().unwrap(), column_types) };
+    pub fn new(column_types: &[Type]) -> Option<TreeStore> {
+        let column_types_ffi: Vec<::glib_ffi::GType> = column_types.iter().map(|n| n.to_glib()).collect();
+        let tmp_pointer = unsafe { ffi::gtk_tree_store_newv(column_types.len().to_i32().unwrap(), column_types_ffi.as_slice()) };
         check_pointer!(tmp_pointer, TreeStore, G_OBJECT_FROM_TREE_STORE)
     }
 
-    pub fn set_column_types(&self, column_types: &[GType]) {
-        unsafe { ffi::gtk_tree_store_set_column_types(self.pointer, column_types.len().to_i32().unwrap(), column_types) }
+    pub fn set_column_types(&self, column_types: &[Type]) {
+        let column_types_ffi: Vec<::glib_ffi::GType> = column_types.iter().map(|n| n.to_glib()).collect();
+        unsafe { ffi::gtk_tree_store_set_column_types(self.pointer, column_types.len().to_i32().unwrap(), column_types_ffi.as_slice()) }
     }
 
     pub fn set_string(&self, iter: &TreeIter, column: i32, text: &str) {
@@ -117,7 +119,7 @@ impl TreeStore {
         }
     }
 
-    pub fn set_value(&self, iter: &TreeIter, column: i32, value: &gtk::GValue) {
+    pub fn set_value(&self, iter: &TreeIter, column: i32, value: &GValue) {
         unsafe { ffi::gtk_tree_store_set_value(self.pointer, iter.unwrap_pointer(), column, value.unwrap_pointer()) }
     }
